requirements.txt
alabaster==0.7.16
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.1
crispy-bootstrap5==2024.2
diff-match-patch==20230430
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-coverage-plugin==3.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-docs==0.3.3
django-extensions==3.2.3
django-extensions-models==1.0
django-filter==24.3
django-health-check==3.18.3
django-import-export==4.1.1
django-redis==5.4.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.2
docutils==0.20.1
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==26.1.0
filelock==3.15.4
flake8==7.1.1
flower==2.0.1
greenlet==3.0.3
gunicorn==22.0.0
humanize==4.10.0
identify==2.6.0
idna==3.7
imagesize==1.4.1
inflection==0.5.1
isort==5.13.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
linkpreview==0.9.0
Markdown==3.6
MarkupSafe==2.1.5
mccabe==0.7.0
mypy==1.11.1
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nodeenv==1.9.1
packaging==24.1
pep8==1.7.1
pillow==10.4.0
platformdirs==4.2.2
pprintpp==0.4.0
pre-commit==3.8.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycodestyle==2.12.1
pydot==3.0.1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
sentry-sdk==2.12.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-django==2.5
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
tablib==3.5.0
tomli==2.0.1
tornado==6.4.1
types-PyYAML==6.0.12.20240724
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
